requirements.txt
python-binance
python-dotenv
pytest
requests
